4/requirements.txt
paddleocr
paddlepaddle
torch
torchvision
transformers
streamlit
pillow
